implement vue single file components
