+++ src/sass/theme.scss
@@ -77,3 +77,5 @@ a {
 @import "../../node_modules/bulma/sass/helpers/_all";
 @import "../../node_modules/bulma/sass/layout/_all";
 @import "../../node_modules/buefy/src/scss/buefy";
+
+@import "./github";
