+++ src/vue/components/heading1.vue
@@ -1,0 +1,16 @@
+<template>
+  <h1
+    :id="id"
+    style="text-transform:uppercase"
+    class="pb-5">
+    <strong><slot></slot></strong>
+  </h1>
+</template>
+
+<script>
+export default {
+  props: [
+    'id',
+  ],
+}
+</script>
