+++ src/vue/components/markdownpreview.vue
@@ -1,0 +1,56 @@
+<template>
+  <div v-if="mdPlaceholder">
+    <b-skeleton
+      width="80%"
+      height="40px"
+      :animated="mdLoading">
+    </b-skeleton>
+    <b-skeleton
+      width="100%"
+      :count="2"
+      :animated="mdLoading">
+    </b-skeleton>
+    <br>
+    <b-skeleton
+      width="100%"
+      :count="3"
+      :animated="mdLoading">
+    </b-skeleton>
+  </div>
+  <div v-else class="markdown-body" v-html="mdPreview">
+    {{ mdPreview }}
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    url: '',
+  },
+  data: function(){
+    return {
+      mdLoading: false,
+      mdPlaceholder: true,
+      mdPreview: '',
+    }
+  },
+  watch: {
+    url: function(){
+      console.log('url changed');
+      this.mdLoading = true;
+      this.mdPlaceholder = true;
+      fetch(`/api/markdown/?url=${this.url}`)
+        .then(res => res.json())
+        .then((data) => {
+            this.mdLoading = false;
+            if(data.markdown != "Data could not be retrieved."){
+              this.mdPlaceholder = false;
+              this.mdPreview = data.markdown;
+              console.log('updating', this.$el);
+              this.$nextTick(() => window.renderMathInElement(this.$el));
+            }
+        })
+    },
+  },
+}
+</script>
