+++ webpack.config.js
@@ -1,0 +1,22 @@
+const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const path = require('path');
+
+module.exports = {
+  mode: 'development',
+  module: {
+    rules: [
+      {
+        test: /\.vue$/,
+        loader: 'vue-loader'
+      }
+    ]
+  },
+  plugins: [
+    new VueLoaderPlugin(),
+  ],
+  resolve:{
+    alias:{
+      'vue$': 'vue/dist/vue.esm.js'
+    }
+  }
+}
